+++ yawast/scanner/plugins/http/servers/php.py
@@ -3,19 +3,23 @@
 #  See the LICENSE file or go to https://yawast.org/license/ for full license details.
 
 from typing import List, cast
+from urllib.parse import urljoin
 
 from packaging import version
 
 from yawast.reporting.enums import Vulnerabilities
+from yawast.scanner.plugins.evidence import Evidence
 from yawast.scanner.plugins.http import version_checker
 from yawast.scanner.plugins.result import Result
+from yawast.scanner.session import Session
+from yawast.shared import network
 
 
 def check_version(banner: str, raw: str, url: str) -> List[Result]:
+    results = []
+
     if not banner.startswith("PHP/"):
         return []
-
-    results = []
 
     # we've got a PHP version
     results.append(
@@ -40,5 +44,28 @@ def check_version(banner: str, raw: str, url: str) -> List[Result]:
                 raw,
             )
         )
+
+    return results
+
+
+def find_phpinfo(session: Session, links: List[str]) -> List[Result]:
+    results = []
+
+    targets = ["phpinfo.php", "info.php", "version.php", "x.php"]
+
+    for link in links:
+        if link.endswith("/"):
+            for target in targets:
+                turl = urljoin(session.url, target)
+                res = network.http_get(turl, False)
+
+                if res.status_code == 200 and '<h1 class="p">PHP Version' in res.text:
+                    results.append(
+                        Result.from_evidence(
+                            Evidence.from_response(res),
+                            f"PHP Info Found: {turl}",
+                            Vulnerabilities.SERVER_PHP_PHPINFO,
+                        )
+                    )
 
     return results